badDend {gtools}	R Documentation

Dataset That Crashes Base:::Plot.Dendogram with 'Node Stack Overflow'

Description

Base:::Plot.Dendogram() will generate a 'Node Stack Overflow' when run on a dendrogram appropriately constructed from this data set.

Usage

data("badDend")

Format

The format is: num [1:2047, 1:12] 1 2 3 4 5 6 7 8 9 10 ... - attr(*, "dimnames")=List of 2 ..$ : NULL ..$ : chr [1:12] "X" "V1" "V2" "V3" ...

Note

See help page for unByteCode to see how to construct the 'bad' dentrogram from this data and how to work around the issue.

Examples

data(badDend)

[Package gtools version 3.8.1 Index]