foldchange {gtools}	R Documentation

Compute fold-change or convert between log-ratio and fold-change.

Description

foldchange computes the fold change for two sets of values. logratio2foldchange converts values from log-ratios to fold changes. foldchange2logratio does the reverse.

Usage

foldchange(num,denom)
logratio2foldchange(logratio, base=2)
foldchange2logratio(foldchange, base=2)

Arguments

num,denom
vector/matrix of numeric values

logratio
vector/matrix of log-ratio values

foldchange
vector/matrix of fold-change values

base
Exponential base for the log-ratio.

Details

Fold changes are commonly used in the biological sciences as a mechanism for comparing the relative size of two measurements. They are computed as: num/denom if num>denom, and as -denom/num otherwise.

Fold-changes have the advantage of ease of interpretation and symmetry about num=denom, but suffer from a discontinuty between -1 and 1, which can cause significant problems when performing data analysis. Consequently statisticians prefer to use log-ratios.

Value

A vector or matrix of the same dimensions as the input containing the converted values.

Author(s)

Gregory R. Warnes greg@warnes.net

Examples

  a <- 1:21
  b <- 21:1

  f <- foldchange(a,b)

  cbind(a,b,f)

[Package gtools version 3.8.1 Index]