gtools {gtools}	R Documentation

gtools: Various R Programming Tools

Description

Functions to assist in R programming, including:

developing, updating, and maintaining R and R packages

'ask', 'checkRVersion', 'getDependencies', 'keywords', 'scat'

calculate the logit and inverse logit transformations

'logit', 'inv.logit'

test if a value is missing, empty, contains only NA and NULL values, or is a 'try-error'

'invalid'

manipulate R's .Last function

'addLast'

define macros

'defmacro', 'strmacro'

detect odd and even integers

'odd', 'even'

convert strings containing non-ASCII characters (like single quotes) to plain ASCII

'ASCIIfy'

perform a binary search

'binsearch'

sort strings containing both numeric and character components

'mixedsort', 'mixedorder'

create a factor variable from the quantiles of a continuous variable

'quantcut'

enumerate permutations and combinations

'combinations', 'permutation'

calculate and convert between fold-change and log-ratio

'foldchange', 'logratio2foldchange', 'foldchange2logratio'

calculate probabilities and generate random numbers from Dirichlet distributions

'rdirichlet', 'ddirichlet'

apply a function over adjacent subsets of a vector

'running'

modify the TCP\_NODELAY ('de-Nagle') flag for socket objects

'tcpNoDelay'

efficient 'rbind' of data frames, even if the column names don't match

'smartbind'

generate significance stars from p-values

'stars.pval'

convert characters to/from ASCII codes

asc, chr

convert character vector to ASCII representation

'ASCIIfy'

apply title capitalization rules to a character vector

'capwords'

[Package gtools version 3.9.2 Index]