gtools-defunct {gtools}	R Documentation

Defunct Functions in package gtools

Description

The functions or variables listed here are no longer part of package gtools.

Details

assert is a defunct synonym for stopifnot.

addLast has been replaced by lastAdd, which has the same purpose but appled using different syntax.

capture and capture.output have been removed in favor of capture.output from the utils package.

See Also

Defunct, stopifnot, lastAdd, capture.output

[Package gtools version 3.8.1 Index]