gtools-deprecated {gtools}	R Documentation

Deprecated Functions in the gtools package

Description

These functions are provided for compatibility with older versions of gtools, and may be defunct as soon as the next release.

Details

gtools currently contains no deprecated functions.

See Also

Deprecated

[Package gtools version 3.8.1 Index]