keywords {gtools}	R Documentation

List valid keywords for R man pages

Description

List valid keywords for R man pages

Usage

keywords(topic)

Arguments

topic
object or man page topic

Details

If topic is provided, return a list of the keywords associated with topic. Otherwise, display the list of valid R keywords from the R doc/KEYWORDS file.

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

help

Examples

## Show all valid R keywords
keywords()

## Show keywords associated with the 'merge' function
keywords(merge)
keywords("merge")

[Package gtools version 3.8.1 Index]