loadedPackages {gtools}	R Documentation

Provide Name, Version, and Path of Loaded Package Namespaces

Description

Provide name, version, and path of loaded package namespaces

Usage

loadedPackages(silent = FALSE)

Arguments

silent
Logical indicating whether the results should be printed

Value

Invisibly returns a data frame containing one row per loaded package namespace, with columns:

Package
Package name

Version
Version string

Path
Path to package files

SearchPath
Either the index of the package namespace in the current search path, or '-' if the package namespace is not in the search path. '1' corresponds to the top of the search path (the first namespace searched for values).

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

loadedNamespaces, packageVersion, search, find.package

Examples

  loadedPackages()

[Package gtools version 3.8.1 Index]