roman2int {gtools}	R Documentation

Convert Roman Numerals to Integers

Description

Convert roman numerals to integers

Usage

roman2int(roman)

Arguments

roman
character vector containing roman numerals

Details

This functon will convert roman numerals to integers without the upper bound imposed by R (3899), ignoring case.

Value

A integer vector with the same length as roman. Character strings which are not valid roman numerals will be converted to NA.

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

as.roman

Examples

roman2int( c('I', 'V', 'X', 'C', 'L', 'D', 'M' )  )

# works regardless of case
roman2int( 'MMXVI' )
roman2int( 'mmxvi' )

# works beyond R's limit of 3899
val.3899 <- 'MMMDCCCXCIX'
val.3900 <- 'MMMCM'
val.4000 <- 'MMMM'
as.numeric(as.roman( val.3899 ))
as.numeric(as.roman( val.3900 ))
as.numeric(as.roman( val.4000 ))

roman2int(val.3899)
roman2int(val.3900)
roman2int(val.4000)

[Package gtools version 3.8.1 Index]