split_path {gtools}	R Documentation

Split a File Path into Components

Description

This function converts a character scalar containing a valid file path into a character vector of path components (e.g. directories).

Usage

split_path(x, depth_first = TRUE)

Arguments

x
character scalar. Path to be processed.

depth_first
logical. Should path be returned depth first? Defaults to TRUE.

Value

Character vector of path components, depth first.

[Package gtools version 3.8.1 Index]