ELISA {gtools}	R Documentation

Data from an ELISA assay

Description

Observed signals and (for some observations) nominal concentrations for samples that were aliquoted to multiple assay plates, which were read multiple times on multiple days.

Usage

data(ELISA)

Format

a data frame with the following columns:

PlateDay

factor. Specifies one of four physically disctinct 96 well plates

Read

factor. The signal was read 3 times for each plate.

Description

character. Indicates contents of sample.

Concentration

numeric. Nominal concentration of standards (NA for all other samples).

Signal

numeric. Assay signal. Specifically, optical density (a colorimetric assay).

Source

Anonymized data.

[Package gtools version 2.6.1 Index]