gtools-deprecated {gtools}	R Documentation

Deprecated Functions in the gtools package

Description

These functions are provided for compatibility with older versions of gtools, and may be defunct as soon as the next release.

Usage

assert(FLAG)

Arguments

FLAG 	 Expression that should evaluate to a boolean vector

Details

The original help page for these functions is often available at help("oldName-deprecated") (note the quotes).

assert is a deprecated synonym for stopifnot.

See Also

Deprecated

[Package gtools version 2.6.1 Index]