quantcut {gtools}	R Documentation

Create a Factor Variable Using the Quantiles of a Continuous Variable

Description

Create a factor variable using the quantiles of a continous variable.

Usage

quantcut(x, q=seq(0,1,by=0.25), na.rm=TRUE, ...)

Arguments

x 	 Continous variable.

q 	 Vector of quantiles used for creating groups. Defaults to seq(0, 1, by=0.25). See quantile for details.

na.rm 	 Boolean indicating whether missing values should be removed when computing quantiles. Defaults to TRUE.

... 	 Optional arguments passed to cut.

Details

This function uses quantile to obtain the specified quantiles of x, then calls cut to create a factor variable using the intervals specified by these quantiles.

It properly handles cases where more than one quantile obtains the same value, as in the second example below. Note that in this case, there will be fewer generated factor levels than the specified number of quantile intervals.

Value

Factor variable with one level for each quantile interval given by q.

Author(s)

Gregory R. Warnes greg@random-technologies-llc.com

See Also

cut, quantile

Examples

  ## create example data
  
  x <- rnorm(1000)

  ## cut into quartiles
  quartiles <- quantcut( x )
  table(quartiles)

  ## cut into deciles
  deciles <- quantcut( x, seq(0,1,by=0.1) )
  table(deciles)

  ## show handling of 'tied' quantiles.
  x <- round(x)  # discretize to create ties
  stem(x)        # display the ties
  deciles <- quantcut( x, seq(0,1,by=0.1) )

  table(deciles) # note that there are only 5 groups (not 10) 
                 # due to duplicates

[Package gtools version 2.6.1 Index]